requirements.txt
asgiref
Django
djangorestframework
pip
sqlparse